# rideshare-rider-service/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rider Service API",
    "description": "Complete API for rider authentication, profile management, and ride requests in a ride-hailing platform.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "dev@swiftride.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1/riders",
      "description": "Local development"
    },
    {
      "url": "https://api.yourapp.com/riders",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Location": {
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TripResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "driverName": {
            "type": "string"
          },
          "vehicle": {
            "type": "object",
            "properties": {
              "model": {
                "type": "string"
              },
              "plate": {
                "type": "string"
              },
              "color": {
                "type": "string"
              }
            }
          },
          "pickup": {
            "$ref": "#/components/schemas/Location"
          },
          "dropoff": {
            "$ref": "#/components/schemas/Location"
          },
          "fare": {
            "type": "number"
          },
          "distance": {
            "type": "number",
            "description": "Distance in kilometers"
          },
          "duration": {
            "type": "number",
            "description": "Duration in minutes"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_MATCHING",
              "MATCHED",
              "ACCEPTED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Invalid or missing token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/riders/me/trips": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get rider's trip history",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of trips",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "trips": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TripResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/riders/me/trips/{tripId}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get detailed information about a specific trip",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}